test: dedupe helpers and refine file access policy checks
